pyproject.toml
[tool.poetry]
name = "mentnl-ingest"
version = "0.1.0"
description = ""
authors = ["jthetzel <jthetzel@gmail.com>"]
readme = "README.md"
packages = [{include = "mentnl_ingest"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.79.0"}
httpx = "^0.23.0"
python-dotenv = "^0.20.0"
google-cloud-firestore = "^2.6.0"
typer = {extras = ["all"], version = "^0.6.1"}
dask = "^2022.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
